@@ -1,10 +1,10 @@
 package controller
 
 import (
-	"red-server/global"
-	"red-server/model"
-	"red-server/service"
-	"red-server/utils"
+	"hot-chat/global"
+	"hot-chat/model"
+	"hot-chat/service"
+	"hot-chat/utils"
 
 	"errors"
 
