@@ -2,7 +2,7 @@
 package starter
 
 import (
-	"red-server/global"
+	"hot-chat/global"
 
 	"github.com/rifflock/lfshook"
 	"github.com/sirupsen/logrus"
@@ -19,17 +19,17 @@ func (s *LogStarter) Name() string {
 func (s *LogStarter) Init() {
 	logger := logrus.StandardLogger()
 	// 定义日志格式
-	formatter := global.CONFIG.Log.GetFormatter()
+	formatter := global.Config.Log.GetFormatter()
 	logger.SetFormatter(formatter)
 
 	// 日志级别
-	level := global.CONFIG.Log.GetLevel()
+	level := global.Config.Log.GetLevel()
 	logger.SetLevel(level)
 
 	// 日志级别分隔
 	writerMap := lfshook.WriterMap{}
 	for _, level := range logrus.AllLevels {
-		writerMap[level] = global.CONFIG.Log.GetOutput(level)
+		writerMap[level] = global.Config.Log.GetOutput(level)
 	}
 	lfHook := lfshook.NewHook(writerMap, formatter)
 	logger.AddHook(lfHook)
