@@ -2,7 +2,7 @@ package service
 
 import (
 	"errors"
-	"red-server/model"
+	"hot-chat/model"
 
 	"github.com/segmentio/ksuid"
 	"github.com/shopspring/decimal"
